port-scanner-fn/requirements.txt
requests==2.32.3
ipaddress==1.0.23
chardet==4.0.0
docker==4.2.2
tldextract==3.4.0
redis==5.0.7
more-itertools==10.3.0
python-dotenv==1.0.1
certifi==2025.1.31
